requirements.txt
 pandas
PyPDF2
pdf2image
opencv-python-headless
numpy
pytesseract
Pillow
argparse
